requirements.txt
exif==1.6.0
ExifRead==3.0.0
pillow==10.2.0
plum-py==0.8.7
